renamed be_var -> config and general -> shared
