+++ example/molbe_io_fcidump.py
@@ -1,7 +1,7 @@
 # Illustrates how fcidump file containing fragment hamiltonian
 # can be generated using be2fcidump
 
-from quemb.general import config
+from quemb.shared import config
 from quemb.molbe import BE, fragpart
 from quemb.molbe.misc import be2fcidump, libint2pyscf
