+++ src/quemb/shared/__init__.py
@@ -1,4 +1,4 @@
-"""This package contains functionality that is relevant for both kbe and molbe.
+"""This package contains functionality that is relevant for both `kbe` and `molbe`.
 
 This can be either shared software infastructure or physical concepts that apply to
 both periodic and non-periodic systems.
